+++ src/LinkedIn.js
@@ -1,6 +1,5 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
-import '../assets/index.css';
 
 const getPopupPositionProperties = ({
   width = 600,
@@ -10,6 +9,18 @@ const getPopupPositionProperties = ({
   const top = (screen.height / 2) - (height / 2);
   return `left=${left},top=${top},width=${width},height=${height}`;
 }
+
+const generateRandomString = (length=20) => {
+  var result           = '';
+  var characters       = 'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789';
+  var charactersLength = characters.length;
+  for ( var i = 0; i < length; i++ ) {
+     result += characters.charAt(Math.floor(Math.random() * charactersLength));
+  }
+  return result;
+}
+
+const LINKEDIN_OAUTH2_STATE = 'linkedin_oauth2_state'
 
 export class LinkedIn extends Component {
   static propTypes = {
@@ -29,8 +40,10 @@ export class LinkedIn extends Component {
   }
 
   getUrl = () => {
-    const { redirectUri, clientId, state, scope, supportIE, redirectPath } = this.props;
+    const { redirectUri, clientId, scope, supportIE, redirectPath } = this.props;
     const scopeParam = `&scope=${supportIE ? scope : encodeURI(scope)}`;
+    const state = generateRandomString();
+    localStorage.setItem(LINKEDIN_OAUTH2_STATE, state);
     const linkedInAuthenLink = `https://www.linkedin.com/oauth/v2/authorization?response_type=code&client_id=${clientId}&redirect_uri=${redirectUri}${scopeParam}&state=${state}`;
     if (supportIE) {
       const redirectLink = `${window.location.origin}${redirectPath}?linkedin_redirect_url=${encodeURIComponent(linkedInAuthenLink)}`;
@@ -40,7 +53,7 @@ export class LinkedIn extends Component {
   }
 
   receiveMessage = (event) => {
-    const { state } = this.props;
+    const state = localStorage.getItem(LINKEDIN_OAUTH2_STATE);
     if (event.origin === window.location.origin) {
       if (event.data.errorMessage && event.data.from === 'Linked In') {
         // Prevent CSRF attack by testing state
@@ -53,6 +66,7 @@ export class LinkedIn extends Component {
       } else if (event.data.code && event.data.from === 'Linked In') {
         // Prevent CSRF attack by testing state
         if (event.data.state !== state) {
+          console.error('State does not match');
           this.popup && this.popup.close();
           return;
         }
@@ -74,7 +88,7 @@ export class LinkedIn extends Component {
 
 
   render() {
-    const { className, disabled, children, renderElement } = this.props;
+    const { className, disabled, children, renderElement, style } = this.props;
     if (renderElement) {
       return renderElement({ onClick: this.handleConnectLinkedInClick, disabled })
     }
@@ -84,19 +98,25 @@ export class LinkedIn extends Component {
         onClick={this.handleConnectLinkedInClick}
         className={className}
         disabled={disabled}
+        style={style ? style : {
+          background: 'none',
+          color: 'inherit',
+          border: 'none',
+          padding: 0,
+          cursor: 'pointer',
+          font: 'inherit',
+          outline: 'inherit',
+        }}
       >
         {children}
       </button>
-
     );
   }
 }
 
 LinkedIn.defaultProps = {
-  className: 'btn-linkedin',
   disabled: false,
   children: (<img src={require('../assets/linkedin.png')} alt="Log in with Linked In" style={{ maxWidth: '180px' }} />),
-  state: 'fdsf78fyds7fm',
   supportIE: false,
   redirectPath: '/linkedin',
   scope: 'r_emailaddress',