+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+## 1.0.9
+### Features
+- Remove prop `state`. It's generated automatically
+- Add `style` prop
+### Fixes
+- Remove `index.css` to fix #13, #30
+
+### Chores
+- Remove default class `btn-linkedin`
+- Inline button style. (TODO: To use children as renderElement in 2.x)
+- Update README to use image from `react-linkedin-login-oauth2/assets/linkedin.png`
+
+
 ## 1.0.8
 
 ### Fixes